+++ lib/sub_page/company/my_company/my_company_widget.dart
@@ -499,8 +499,9 @@ class _MyCompanyWidgetState extends State<MyCompanyWidget>
                                         'ReviewList',
                                         queryParameters: {
                                           'companyID': serializeParam(
-                                            myCompanyVarItem.companyID,
-                                            ParamType.int,
+                                            myCompanyVarItem.companyID
+                                                .toDouble(),
+                                            ParamType.double,
                                           ),
                                         }.withoutNulls,
                                         extra: <String, dynamic>{
